Prevent page break between commit heading and body
Currently every commit entry in the document starts like:
      <Heading> <Stats> <Body>

Sometimes the page breaks between the commit's heading and its body.
That looks bad.

Conventional methods to prevent the page breaking did not work. I tried
using 'nowidow' package and setting the penalties like `\widowpenalty`.

I think the reason of their failure is due to <Stats> block being inside
verbatim environment. So, I have to use more low-level methods to solve
the problem.

My next guess is to wrap `\subsection` with `\renewcommand` so that
to add `\Needspace` right before every subsection starts. However,
this causes an error with the table of contents when compiling the
document.

I need a more tricky solution to prepent `\needspace` to headings in
the text but not to change the behaviour of headings inside the table
of contents.

To do so I use `\pretocmd` for the subsection's addtocontents command
in such a way to preserve things that `\addtocontents{toc}` achieves
to use as text inside the "contents" section (the fisrt argument of
addtocontents) and prepend `\Needspace{}` to the text used as section
heading inside the document's body (the third argument of
addtocontents).

The solution is inspired with Gonzalo Medina's answer at:

  https://tex.stackexchange.com/questions/56546